Move pytest fixtures to a central location.
